+++ src/python/pants/base/exceptions.py
@@ -30,14 +30,6 @@ class MappingError(Exception):
     """Indicates an error mapping addressable objects."""
 
 
-class UnaddressableObjectError(MappingError):
-    """Indicates an un-addressable object was found at the top level."""
-
-
-class DuplicateNameError(MappingError):
-    """Indicates more than one top-level object was found with the same name."""
-
-
 class ResolveError(MappingError):
     """Indicates an error resolving targets."""
