--- src/python/pants/console/README.md
@@ -1,3 +1,0 @@
-/src/python/pants/console/
-
-Common code useful for REPL consoles.
