--- src/python/pants/fs/README.md
@@ -1,3 +1,0 @@
-/src/python/pants/fs/
-
-Common code useful for working with files.
