+++ src/python/pants/init/global_subsystems.py
@@ -2,7 +2,6 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 from pants.goal.run_tracker import RunTracker
-from pants.process.subprocess import Subprocess
 from pants.reporting.reporting import Reporting
 from pants.scm.subsystems.changed import Changed
 
@@ -11,4 +10,4 @@ class GlobalSubsystems:
     @classmethod
     def get(cls):
         """Subsystems used outside of any task."""
-        return {Reporting, RunTracker, Changed, Subprocess.Factory}
+        return {Reporting, RunTracker, Changed}
