--- src/python/pants/java/README.md
@@ -1,3 +1,0 @@
-/src/python/pants/java/
-
-Logic to launch Java tools on Pants' behalf.
