--- tests/python/pants_test/fs/BUILD
@@ -1,6 +1,0 @@
-# Copyright 2020 Pants project contributors (see CONTRIBUTORS.md).
-# Licensed under the Apache License, Version 2.0 (see LICENSE).
-
-python_tests(name='tests')
-
-python_library()
